requirements.txt
dj-database-url==0.5.0
Django
django-bootstrap4
django-crispy-forms==1.7.2
django-email-extras==0.3.3
django-filter==2.0.0
django-forms-builder==0.14.0
django-password-reset==1.0
git+https://github.com/fellipeh/django-tenant-schemas
django-tools==0.40.4
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
future==0.15.0
lxml==4.2.4
MarkupSafe==1.0
MonthDelta==0.9.1
Pillow==5.2.0
psycopg2==2.7.5
python-decouple==3.1
python-gnupg==0.4.3
pytz==2018.3
sphinx-me==0.3
Unidecode==1.0.22
